requirements.txt
feedparser
requests>=2.31.0
beautifulsoup4>=4.12.0
Pillow>=10.0.0
google-genai>=1.20.0
groq>=0.9.0
huggingface-hub>=0.20.0
python-dateutil>=2.8.2
